requirements.txt
Pillow==5.3.0
pyFirmata==1.0.3
pyserial==3.4
